# docs/api/openapi/session-service.yaml
openapi: 3.1.0
info:
  title: Noema Session Service API
  version: 0.1.0
  description: |
    Session Service manages learning session lifecycle, card queuing,
    attempt recording, and session state transitions.

    **Bounded Context:** Learning Sessions
    **Database:** PostgreSQL + Redis
    **Events:** 12 session domain events via Redis Streams

    All mutation responses include `agentHints` for downstream agent consumption.
  contact:
    name: Noema Platform
  license:
    name: Proprietary
servers:
  - url: http://localhost:3003
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==========================================================================
  # Schemas
  # ==========================================================================
  schemas:
...
    SessionState:
      type: string
      enum: [active, paused, completed, abandoned, expired]
...
    LearningMode:
      type: string
      enum: [exploration, goal_driven, exam_oriented, synthesis]
...
    StartSessionInput:
      type: object
      properties:
        deckQueryId:
          $ref: "#/components/schemas/DeckQueryLogId"
        learningMode:
          $ref: "#/components/schemas/LearningMode"
        teachingApproach:
          type: string
          default: standard
        schedulingAlgorithm:
          type: string
          default: fsrs
        loadoutId:
          type: string
        loadoutArchetype:
          type: string
        config:
          $ref: "#/components/schemas/SessionConfig"
        initialCardIds:
          type: array
          items:
            $ref: "#/components/schemas/CardId"
          description: Pre-resolved card IDs for the initial queue
      required:
        - deckQueryId
        - learningMode
        - config
        - initialCardIds
...
paths:
...
  /v1/sessions:
    post:
      operationId: startSession
      tags: [Sessions]
      summary: Start a new learning session
      description: |
        Creates a new session with FSM state=active, initializes the card queue,
        and publishes `session.started`. Requires pre-resolved cardIds
        (the Learning Agent resolves the DeckQuery before calling this).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSessionInput"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Session"
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "422":
          description: Business rule violation

    get:
      operationId: listSessions
      tags: [Sessions]
      summary: List sessions for authenticated user
      parameters:
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/SessionState"
        - name: learningMode
          in: query
          schema:
            $ref: "#/components/schemas/LearningMode"
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated session list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      sessions:
                        type: array
                        items:
                          $ref: "#/components/schemas/Session"
                      total:
                        type: integer
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "401":
          description: Unauthorized
  /v1/sessions/{sessionId}:
    get:
      operationId: getSession
      tags: [Sessions]
      summary: Get session by ID
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
      responses:
        "200":
          description: Session found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Session"
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "401":
          description: Unauthorized
        "404":
          description: Session not found
  /v1/sessions/{sessionId}/pause:
    post:
      operationId: pauseSession
      tags: [Sessions]
      summary: Pause an active session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Session paused
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Session"
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "400":
          description: Invalid state transition
        "401":
          description: Unauthorized
        "404":
          description: Session not found
  /v1/sessions/{sessionId}/resume:
    post:
      operationId: resumeSession
      tags: [Sessions]
      summary: Resume a paused session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
      responses:
        "200":
          description: Session resumed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Session"
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "400":
          description: Invalid state transition
        "401":
          description: Unauthorized
        "404":
          description: Session not found
  /v1/sessions/{sessionId}/complete:
    post:
      operationId: completeSession
      tags: [Sessions]
      summary: Complete an active session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
      responses:
        "200":
          description: Session completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Session"
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "400":
          description: Invalid state transition
        "401":
          description: Unauthorized
        "404":
          description: Session not found
  /v1/sessions/{sessionId}/abandon:
    post:
      operationId: abandonSession
      tags: [Sessions]
      summary: Abandon an active or paused session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Session abandoned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Session"
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "400":
          description: Invalid state transition
        "401":
          description: Unauthorized
        "404":
          description: Session not found

  # ---------- Attempts ----------
  /v1/sessions/{sessionId}/attempts:
    post:
      operationId: recordAttempt
      tags: [Attempts]
      summary: Record a card review attempt
      description: |
        Records a single attempt within an active session.
        This is the most critical write operation in Noema — it publishes
        `attempt.recorded` which is consumed by 5+ services.
        Automatically updates session stats and queue item status.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordAttemptInput"
      responses:
        "201":
          description: Attempt recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Attempt"
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Session not found
        "422":
          description: Session not active

    get:
      operationId: listAttempts
      tags: [Attempts]
      summary: List attempts for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Paginated attempt list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      attempts:
                        type: array
                        items:
                          $ref: "#/components/schemas/Attempt"
                      total:
                        type: integer
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "401":
          description: Unauthorized
        "404":
          description: Session not found
  /v1/sessions/{sessionId}/attempts/{attemptId}/hint:
    post:
      operationId: requestHint
      tags: [Attempts]
      summary: Request a progressive hint for current attempt
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
        - name: attemptId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AttemptId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestHintInput"
      responses:
        "200":
          description: Hint request recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      acknowledged:
                        type: boolean
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "401":
          description: Unauthorized
        "404":
          description: Session or attempt not found

  # ---------- Queue Management ----------
  /v1/sessions/{sessionId}/queue:
    get:
      operationId: getQueue
      tags: [Queue]
      summary: Get current session queue state
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
      responses:
        "200":
          description: Queue items
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SessionQueueItem"
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "401":
          description: Unauthorized
        "404":
          description: Session not found
  /v1/sessions/{sessionId}/queue/inject:
    post:
      operationId: injectQueueItem
      tags: [Queue]
      summary: Inject a card into the session queue
      description: |
        Used by the Learning Agent to dynamically add cards (e.g., remediation
        cards) into an active session's queue at a specific position.
        Publishes `session.queue.injected`.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InjectQueueInput"
      responses:
        "201":
          description: Card injected
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/SessionQueueItem"
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Session not found
        "422":
          description: Session not active
  /v1/sessions/{sessionId}/queue/remove:
    post:
      operationId: removeQueueItem
      tags: [Queue]
      summary: Remove a card from the session queue
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveQueueInput"
      responses:
        "200":
          description: Card removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      removed:
                        type: boolean
                  agentHints:
                    $ref: "#/components/schemas/AgentHints"
                  metadata:
                    $ref: "#/components/schemas/ApiMetadata"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Session or card not found
  /v1/sessions/blueprint/validate:
    post:
      operationId: validateSessionBlueprint
      tags: [Sessions]
      summary: Validate agent-generated session blueprint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                blueprint:
                  type: object
              required: [blueprint]
      responses:
        "200":
          description: Blueprint validation result
        "400":
          description: Validation error
        "401":
          description: Unauthorized
...